mmc: fix the workqueue overload battery waste
After switching to the smdk4412 kernel I noticed a workqueue overload,
battery lasting less, and Android OS showing up at the top as result.

Tracing showed that mmc_host_clk_gate_work was one of the runqueues
called regularly in the background.

The #ifdef is clearly reversed and that leads to the ndelay being called
every time, while ndelay was never called before.

Not yet sure if this fixes the "apparent" regression but it's not bad so
far, after 1h of uptime android OS is below Android System.

Change-Id: Ib908b0c20c4bc8dd889d0d228ee3ea63820b98cf